src/lib/search/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pinecone
openai>=0.27.0
python-dotenv>=0.19.0 